requirements.txt
absl-py==0.8.0
astor==0.8.0
astroid==2.2.5
autopep8==1.4.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
firebase-admin==2.18.0
Flask==1.1.1
Flask-API==1.1
gast==0.3.2
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-pasta==0.1.7
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.3.0
h5py==2.10.0
httplib2==0.13.1
idna==2.8
imageio==2.5.0
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==0.23
mccabe==0.6.1
msgpack==0.6.1
numpy==1.17.2
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
